+++ services/payments/src/simcore_service_payments/api/rest/_dependencies.py
@@ -4,8 +4,8 @@ from typing import Annotated, cast
 
 from fastapi import Depends, FastAPI, Request
 from fastapi.security import OAuth2PasswordBearer
+from servicelib.fastapi.app_state import AppStateMixin
 from servicelib.fastapi.dependencies import get_app, get_reverse_url_mapper
-from servicelib.fastapi.http_client import AppStateMixin
 from sqlalchemy.ext.asyncio import AsyncEngine
 
 from ..._meta import API_VTAG
