+++ services/payments/src/simcore_service_payments/services/resource_usage_tracker.py
@@ -19,7 +19,8 @@ from models_library.products import ProductName
 from models_library.resource_tracker import CreditTransactionId
 from models_library.users import UserID
 from models_library.wallets import WalletID
-from servicelib.fastapi.http_client import AppStateMixin, BaseHttpApi
+from servicelib.fastapi.app_state import AppStateMixin
+from servicelib.fastapi.http_client import BaseHttpApi
 
 from ..core.settings import ApplicationSettings
